dom/bindings/test/test_InstanceOf.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=748983
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 748983</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=748983">Mozilla Bug 748983</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 748983 **/

SimpleTest.waitForExplicitFinish();

function runTest() {
    ok(document instanceof EventTarget, "document is an event target");
    ok(new XMLHttpRequest() instanceof XMLHttpRequest, "instanceof should work on XHR");
    ok(HTMLElement.prototype instanceof Node, "instanceof needs to walk the prototype chain");

    var otherWin = document.getElementById("testFrame").contentWindow;

    ok(otherWin.HTMLElement.prototype instanceof otherWin.Node, "Same-origin instanceof of a interface prototype object should work, even if called cross-origin");
    ok(!(otherWin.HTMLElement.prototype instanceof Node), "Cross-origin instanceof of a interface prototype object shouldn't work");

    // We need to reset HTMLElement.prototype.__proto__ to the original value
    // before using anything from the harness, otherwise the harness code breaks
    // in weird ways.
    HTMLElement.prototype.__proto__ = otherWin.Element.prototype;
    var [ shouldSucceed, shouldFail ] = otherWin.runTest();
    shouldSucceed = shouldSucceed && HTMLElement.prototype instanceof otherWin.Element;
    shouldFail = shouldFail && HTMLElement.prototype instanceof Element;
    HTMLElement.prototype.__proto__ = Element.prototype;

    ok(shouldSucceed, "If an interface prototype object is on the protochain then instanceof with the interface object should succeed");
    ok(!shouldFail, "If an interface prototype object is not on the protochain then instanceof with the interface object should succeed");

    SimpleTest.finish();
}

</script>
</pre>
<iframe id="testFrame" src="file_InstanceOf.html" onload="runTest()"></iframe>
</body>
</html>
